+++ src/event/event.ts
@@ -1,86 +1,9 @@
 import type http from 'http'
-import type {
-  IncomingMessage,
-  ServerResponse,
-  CompatibilityEventHandler,
-  EventHandler,
-  H3EventContext,
-  HandlerResponse,
-  LazyEventHandler
-} from '../types'
-import { callHandler } from '../handler'
+import type { IncomingMessage, ServerResponse, H3EventContext } from '../types'
 import { MIMES } from '../utils'
 import { H3Response } from './response'
 
-export function defineEventHandler <T = any> (handler: EventHandler<T>): EventHandler<T> {
-  handler.__is_handler__ = true
-  return handler
-}
-export const eventHandler = defineEventHandler
-
-export function defineLazyEventHandler (factory: LazyEventHandler): EventHandler {
-  let _promise: Promise<EventHandler>
-  let _resolved: EventHandler
-  const resolveHandler = () => {
-    if (_resolved) { return Promise.resolve(_resolved) }
-    if (!_promise) {
-      _promise = Promise.resolve(factory()).then((r: any) => {
-        const handler = r.default || r
-        if (typeof handler !== 'function') {
-          throw new (TypeError as any)('Invalid lazy handler result. It should be a function:', handler)
-        }
-        _resolved = toEventHandler(r.default || r)
-        return _resolved
-      })
-    }
-    return _promise
-  }
-  return eventHandler((event) => {
-    if (_resolved) {
-      return _resolved(event)
-    }
-    return resolveHandler().then(handler => handler(event))
-  })
-}
-export const lazyEventHandler = defineLazyEventHandler
-
-export interface DynamicEventHandler extends EventHandler {
-  set: (handler: EventHandler) => void
-}
-export function dynamicEventHandler (initial?: EventHandler): DynamicEventHandler {
-  let current: EventHandler | undefined = initial
-  const wrapper = eventHandler((event) => {
-    if (current) {
-      return current(event)
-    }
-  }) as DynamicEventHandler
-  wrapper.set = (handler) => { current = handler }
-  return wrapper
-}
-
-export function isEventHandler (input: any): input is EventHandler {
-  return '__is_handler__' in input
-}
-
-export function toEventHandler (handler: CompatibilityEventHandler): EventHandler {
-  if (isEventHandler(handler)) {
-    return handler
-  }
-  if (typeof handler !== 'function') {
-    throw new (TypeError as any)('Invalid handler. It should be a function:', handler)
-  }
-  return eventHandler((event) => {
-    return callHandler(handler, event.req as IncomingMessage, event.res) as HandlerResponse
-  })
-}
-
-export function createEvent (req: http.IncomingMessage, res: http.ServerResponse): H3Event {
-  return new H3Event(req, res)
-}
-
-type PartialFetchEvent = Pick<FetchEvent, 'respondWith'>
-
-export class H3Event implements PartialFetchEvent {
+export class H3Event implements Pick<FetchEvent, 'respondWith'> {
   '__is_event__' = true
   req: IncomingMessage
   res: ServerResponse
@@ -149,4 +72,8 @@ export class H3Event implements PartialFetchEvent {
 
 export function isEvent (input: any): input is H3Event {
   return '__is_event__' in input
+}
+
+export function createEvent (req: http.IncomingMessage, res: http.ServerResponse): H3Event {
+  return new H3Event(req, res)
 }
